[REF] website_mail_channel: adapt discussion group to actual snippet
Adapt to current snippet/widget guidelines/hierarchy.

task-2152191
