+++ addons/website/static/src/img/snippets_thumbs/s_channel.png
