--- addons/website_mail_channel/views/snippets.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-<template id="remove_external_snippets" inherit_id="website.external_snippets">
-    <xpath expr="//t[@t-install='website_mail_channel']" position="replace"/>
-</template>
-
-<template id="subscribe" inherit_id="website.snippets" name="Snippet Subscribe">
-    <xpath expr="//div[@id='snippet_content']//t[@t-snippet][last()]" position="after">
-        <div>
-            <div class="oe_snippet_thumbnail">
-                <img class="oe_snippet_thumbnail_img" src="/website/static/src/img/snippets_thumbs/s_button_channel_subscribe.png" alt="Subscribe"/>
-                <span class="oe_snippet_thumbnail_title">Discussion Group</span>
-            </div>
-
-            <div class="oe_snippet_body js_follow_alias"
-                      data-id="0"
-                      data-object="mail.channel"
-                      data-follow="off" data-name="Discussion Group">
-                <div class="input-group js_mg_follow_form">
-                    <input
-                          type="email"
-                          name="email"
-                          class="js_follow_email form-control"
-                          placeholder="your email..."/>
-                    <span class="input-group-append">
-                        <button href="#" class="btn btn-primary js_follow_btn">Subscribe</button>
-                    </span>
-                </div>
-                <p class="js_mg_details d-none">
-                    <span class="js_mg_email d-none"><a href="#" class="js_mg_email"><i class="fa fa-envelope-o"/> send mail</a> - </span>
-                    <a href="#" class="js_mg_link"><i class="fa fa-file-o"/> archives</a> -
-                    <a role="button" href="#" class="js_unfollow_btn"><i class="fa fa-times"/> unsubscribe</a>
-                </p>
-                <p class="js_mg_confirmation d-none">
-                    a confirmation email has been sent.
-                </p>
-            </div>
-        </div>
-    </xpath>
-
-
-    <xpath expr="//div[@id='snippet_options']" position="inside">
-        <div data-js='subscribe'
-             data-selector=".js_follow_alias"
-             data-drop-near="p, h1, h2, h3, blockquote, .card">
-            <we-row>
-                <we-select class="select_discussion_list" data-attribute-name="id" data-no-preview="true">
-                    <!-- 'we-button' added programmatically with DB data -->
-                </we-select>
-                <we-button class="fa fa-fw fa-plus" title="Create a public discussion group in your backend"
-                           data-create-channel="" data-no-preview="true"/>
-            </we-row>
-        </div>
-    </xpath>
-
-</template>
-</odoo>
